Added marking and unmarking task functions
Currently, the bot is only able to store the list of task being done but
there is no way of allowing users to track if task is done

Added marking symbols with "X" between "[ ]" braces after the indexing
of task to allow user to track if task is done. User is able to ivoke
markTask() and unmarkTask() functions through entering keywords such as
"mark" and "unmark" respectively followed by the index of the task in
list() function to mark and unmark the tasks.