Add increment Level-2
Following the instructions for Week 3, Level-2 of Mong allows users to
add items into a list, which can be printed in index order when the
"list" command is entered.

Add items to the list: add items from the command line into a list that
displays when the command "list" is typed out.

Index items in the list: items in the list start from the number "1".