Add increment Level-3
Following the instructions for Week 3, Level-3 of Mong introduces a
marking command for users to set items in the command list as completed
or uncompleted.

Create a class of ListItem: use a class to implement the adding,
indexing, marking, and unmarking of each item in the list.

The class will handle the instantiation of each ListItem and setting of
the "isCompleted" variable that is assigned to each ListItem in the
array of ListItems.