Fixed bugs with mark and unmark functions
The code is only able to mark and unmark indexes up to 9 due to the way
the function is implemented.

The number of task that the bot is implemented for is up to 100 meaning
the indexes goes up to 99. Thus it needs to be changed to accomodate
this.

The code is rectified and fixed to fit the constrain.