Integrate mark and unmark functionality into Echo class for users to use
Echo does not have the ability to mark and unmark tasks.

Let's connect the methods from Task Manage to TaskList to Echo classes so that users can set tasks as done and undone.

This ensures users can easily check the list of task if it is done or not done.