Add different types of Task for Legin
Currently, we are only able to add tasks through the bot to the list and
this is not very effective as there may be differing types of tasks that
require more specific details such as due dates or event timings

The bot now is able to have 3 types of Task catergorisation namely,
Todo, Deadline Task and Event. Users need to input in "todo", "deadline"
and "event" keywords at the start of the input before typing the task to
categorise them. Deadline tasks are able to set a deadline for the task
by adding a deadline after the "/by" keyword in the input after the task
and for events it is able to give a range of time with the "/from" and
"/to" keywords after the input of the task.