Add increment Level-4
Level-4 of Mong supports the tracking of three types of data, namely
Todos, Deadlines, and Events. To implement these changes, we use
Inheritance and Polymorphism.

Enumerators for TaskType: we decipher the first word in the command line
as the TaskType. Currently, Mong supports LIST, BYE, MARK, UNMARK,
EVENT, TODO and DEADLINE as valid task types.

Improve code quality: we improve code quality through the use of
constants (rather than magic numbers), and following of coding
standards. For example, indentations of switch statements are in-line
and the following of naming standards.

Inheritance and polymorphism: Event, Todo, and Deadlines are child
classes of the Task class. The toString() function takes on a
different form based on the whether the object is a Event, Todo, or
Deadline object.