Extract command methods from main
The code for the execution of each user command is within a single
switch statement in main().

This makes it difficult to read and trace through the code.

Let's extract the code handling the command executions into separate
methods. This improves the code quality, and also ensures that there is
a single level of abstraction in main() and for the command handling
code.