Extract greet and command receiving methods from main
The code for greeting the user and receiving user input is written in
one block in main(). This makes the code difficult to read and debug.

Let's:
* extract the code for greeting the user to a greetUser() method
* extract the code for receiving user input into a getCommands() method

This allows for code within main and each of the new methods to be at a
single level of abstraction, increasing code quality and readability.