Refactor task adding functions into one method
The addition of tasks to the task list is done by multiple methods, one
each for the Task class or its children. This results in code
duplication, and also terms like "task[taskCount - 1]" which are
difficult to understand.

Let's consolidate the adding of tasks to the task list into a single
method. This reduces code duplication and makes the code easier to
debug.