Modify Task class
Change access modifiers for attributes from private to protected so that subclasses inheriting from Task class can access these attributes. Add extra methods defined in the subclasses so that the subclasses can override them. This is useful when the subclass objects are being referenced/type-casted as a superclass object. By adding these extra methods in the superclass and overridding them in the subclass, it ensures that the overridden methods from the subclass will be invoked instead of the one in the superclass, otherwise the methods found only in the subclass can no longer be invoked, leading to an error.