Add Level-4 increment: Todo, Event, Deadline
Modify Atom.java to allow user to specify, store and track 3 types of tasks, namely todo, deadline and event. By specifying the type of task before adding it to the list, the list now displays more important info, like the type of task, the deadline of the task, the timeline of the event, etc. depending on the type of task. In order to achieve this, the contents of the printList() method are also changed. Constants are declared at the top and used whenever applicable so that they make sense, resulting in improved code readability, but most importantly it gets rid of any magic numbers.