Update and refactor code to improve code quality
Firstly, by adopting the rule of "make the happy path prominent", I managed to simplify some of the if blocks in the code by switching the order of the logic and adding guard clauses. This helped to remove the redundant else block, making the if block a lot neater and easier to read. Secondly, my initial iteration of the code encountered an issue of deep nesting, aka the arrowhead style code, and this was the biggest (visual) flaw in my code. By refactoring my code, I managed to extract the contents in the large if block into 2 separate methods. This resulted in a short, concise and neat code which is still easy to understand and read. Lastly, I also refactored the "magic numbers" as constants, declared at the top of the code. This gives meaning to these arbituary values that I used and helps other readers make sense of them when they read through my code. Next, I also added exception handling functionalities for both the "deadline" and "event" commands in the event that the user enters an incorrect command format. Lastly, I created a method to print out the default exception message.