Use TextUI testing to semi automate the testing process and check for potential regressions
Using the input/output redirection technique, the input.txt stores all the input that are to be fed to the program/source code, and the output will be stored in another txt file called "ACTUAL.txt". The latter is then compared with EXPECTED.txt - a file containing the expected output of the code, given the inputs defined in the input.txt file, and any difference in the files will be displayed in the terminal.

Initially, even after following all the steps correctly and ensuring that the contents in both EXPECTED.txt and ACTUAL.txt LOOK exactly the same, the file compare (FC) command still showed a difference in the 2 files. To resolve this, I had to open these 2 files in my Notepad++ app and highlight all the contents of the file. Narrowing down my search to the lines where the errors were reported, this was when i realised that the difference was actually empty space characters, which cannot be seen by the naked eye. Following that, I went back to my code to make the necessary changes, and now the files are identical!

In summary, this exercise taught me how to semi automate the testing of textUIs using the I/O redirection technique. This tool/skill proved to be effective as it helps to check for potential bugs/regressions whenever we made changes to the code, and this makes the whole testing process easier, instead of having to manually write the test cases.