Rename command execution methods
The methods for executing the user commands have names beginning with
"executeCommand". This makes the method names difficult to read,
and potentially misleading (as one could think that the methods only
executed the commands)

Let's rename these methods with names that reflect what is actually
done when executing these commands. This way, the method names would be
self-explanatory and more readable.