Rename user command method parameters
The methods for the user commands share the same parameter
"commandBody". This is vague and does not indicate what the parameter
does in the method.

Let's give the parameters different names, that align with what the
parameter is used for in the methods. This makes for more readable
code, and for the parameter's use to be self-explantory.