Appal.java and Parser.java: Improve code readability
Appal.java: Added more named constants, as well as abstracted out the
method printMessage() as it appears quite often in the code.

Parser.java: Added a named constant to make the start of the code more
understandable, and added more details of how the Parser extracts
parameters in user input in the comments.