Add Exception classes and try catch statements
Currently, the bot is unable to verify or validate the input of the
users which may lead to various unexpected outcomes and behaviours that
is not intended.

The bot thus added try catch statements for customised exceptions for
the bot. The exceptions are as follows:
1. When user inputs no task for his/her todos, events, deadlines
2. When user inputs an invalid command
3. When user missed out parameters for a given event or deadline
The bot will catch these human errors and output them to the user so
he/her can edit the input and give the bot the expected one.