Appal.java: Add exception and update named constants and handleInput()
Added the ArrayIndexOutOfBoundsException to markTask() as any input that
is less than or equals to zero, or more than 100, will result in this
exception.

Some named constants were updated to make the chatbot interface look
more consistent.

In handleInput(), instead of creating a new Scanner object each time we
take in user input, a Scanner object is instantiated as an attribute of
the Appal class.