Add error handling to handle empty input fields for tasks
Program allows users to successfully log tasks with empty descriptions,
deadlines, start times or end times.

This is bad as it can potentially allow for tasks with empty fields to
be logged which may confuse the user later on.

Let's create a EmptyFieldException class to flag empty deadline/
start date/end date inputs. handleTodo, handleDeadline and handleEvent
methods will throw a EmptyFieldException upon detecting empty field
inputs, which triggers the program to print an error message informing
the user not to input empty fields.

A new EmptyFieldException class was used to clearly signal to future
code readers that the code is meant to handle errors regarding empty
fields.