Add custom exception class to handle invalid formatting of the EVENT command
The implementation of this is quite similar to that of DEADLINE. However, for the EVENT command, there are now 2 keywords to check for now (i.e "/from" and '/to") instead of just 1. Therefore, this exception class handles incorrect inputs such as "from" or "to" without the "/", and empty FROM or TO date/time inputs.