Update Atom to handle errors like incorrect user inputs
Using all the custom exception classes defined earlier, Atom is now able to handle errors and exceptions like incorrect user inputs. For the exception handling, a "try-catch" block, together with the "throw" keyword is implemented for every command block in Atom, so that every command block can throw and catch the corresponding exception depending on the user input. Multiple catch blocks are used if the exception handler in the command block needs to catch and handle different types of exception.

Furthermore, to improve on code readability, I refactored the code by extracting 3 different methods from the if-else command blocks in the main method, namely "addTodoTask", "addDeadlineTask", and "addEventTask". This makes the main part of the code easier to read, understand, manage and it also looks more visually pleasing. I also refactored the size of the task list from "100" to "MAX_NUMBER_OF_TASKS" so as to avoid using magic numbers.