Organise classes into packages
Created 3 new packages.

1. "atom" --> main package/folder storing all the classes and packages related to Atom
2. "exception" --> stores all of Atom's custom exception classes
3. "task" --> stores all the task related classes (i.e Task, Todo, Deadline, Event)

Following that, the package statment is included at the top of every source file. This was achieved easily thanks to the refactoring feature in the IDE. In addition, in the main Atom source code, all the classes residing in the other packages need to be imported so that they can be used in the main Atom source code. I also ensured that all these new additions conformed to the coding standards, like ensuring all the packages' name are in all lower case, and when importing all the classes used in a package, they have to be stated explicitly instead of using the * operator.