Replace ListAddTask() method in Nell with addTask() method in TaskList
The addition of tasks to the task list is done using the ListAddTask()
method, which then calls the addTask() method. This leads to
unnecessary code duplication.

Let's replace the calls to ListAddTask() in Nell with calls to
addTask(), and remove the ListAddTask() method entirely. This makes for
reduced code duplication and more readable code.