Move message printing from Nell to addTask() method in TaskList
The methods to add various task types in Nell print out the same
messages, and only differ in the types of task they add. This results
in unnecessary code duplication

Let's move the message printing code from these methods to the addTask()
method in TaskList. This allows for reduced code duplication and more
readable code.