Amend data format in data.txt
The tasks in data.txt are stored with some attributes in square
brackets, and others in parentheses, for readability. This makes
data.txt difficult to parse during initialisation.

Let's store the attributes in lines, separated by pipes.

This makes data.txt easier to parse, allowing it to be read into the
program upon startup.