Change TaskList to store tasks using ArrayList instead of array
TaskList uses a fixed-sized array to store tasks.

This limits the number of tasks we can store in our list and is
inefficient in terms of memory usage as we cannot dynamically size
the array.

Let's replace the array of tasks with an ArrayList to store the tasks.

This lets us store and unlimited number of tasks and allows more
efficient use of memory through dynamic sizing of the ArrayList.