Add delete feature to Legin
Currently, the bot is unable to delete the task which is is required by
the user when he/she wants to remove a task that is no longer relevant.
The current implementation is of a fixed size array of 100 which makes
it inconvininet to delete task efficiently.

The bot now runs with an ArrayList instead of fixed length array to
leverage on the remove function of the collection. This allows easy
deletion of tasks. The user can use the delete function with the
following input 'delete [task number]'.