Add getAllTaskDetails method to TaskList class
Program has no means of retrieving all tasks logged in TaskList,
making writing the list of tasks to a new file difficult.

Let's create a method getAllTaskDetails that compiles the details of
all logged tasks and returns it as a single string.

This allows the program to write all tasks onto an external file
easily.