Add methods to create and update save file
Program does not save tasks logged into it, meaning all data is lost
once the program is closed.

Let's create a FileReadWriter class with methods to create a new file
to store the logged tasks and write the list of all logged tasks to it
whenever the list of tasks is updated.

Feature of reading from the save file is a work in progress.