Implement Read Task list from save file feature
Program saves list of tasks to external file but does not read from
the file to access the saved task list.

Let's add an initTaskListFromSaveFile method in FileReadWriter class
which populates TaskList according to the list of tasks logged in
the save file.

This allows the program to initialize its list of tasks according to
the list of tasks as of the program's last closure. FileReadWriter has
also been added to a new savemanager package for better organization of
files.