Remove static variable "taskCount" and its methods in Task class
Since Atom is now using an ArrayList instead of an Array to store the Tasks, the built-in method .size() of the ArrayList class can be used to get the size/number of tasks in the list instead. This removes the need for an extra static variable and getter method to read and query the number of current tasks in the list. Doing so makes the code a lot neater, easier and more straightforward too.