Update printList() method
The if condition that checks whether the list is empty was initially written in the main() method. This if block was moved to the printList() method so that the code in the main() looks more concise. Furthermore, it makes more sense to put it in the printList() method as it should first perform a check to ensure that the list is not empty before actually printing the contents of the list.