Update display format for DEADLINE and EVENT tasks after "mark" and "unmark" commands
Initially, after marking/unmarking a DEADLINE or EVENT task, the output will show the type of task and the name of the task, without any additional information. After the change, the output now not shows the type of task and the name of the task, it also shows the date and time the task has to be completed by for a DEADLINE task, or the timeline or duration of the event for an EVENT task.