Add Level-6 increment: Delete
Atom is now able to delete tasks from the list that are no longer needed. A method deleteTask() is created to delete the specified task from the list and print a message to the user, showing the user the deleted task and the number of tasks remaining in the list. In the main() method, similar to the "mark" and "unmark" commands, all the exceptions like incorrect format and missing task id are handled properly and gracefully, with the appropriate error messages. In addition, as mentioned before, all instances of the static variable "taskCount" are now replaced with tasksList.size().