Edit methods for adding a TODO, DEADLINE and EVENT tasks
Initially, the adding of a TODO, DEADLINE and EVENT task creates and adds the specific task object, based on its type to the list, and prints out a "task added" message all in the same method. With the implementation of the Level-7 increment: Save, these methods need to be reused in order to add the task's data stored in the file to the list. However, as the tasks are added from the txt file to the list at the start of the program, all the corresponding task messages will also be printed at the start, flooding the console terminal with loads of "task added" messages, which are not required to be shown to the user. Therefore, I have further refactored the method, essentially separating it into 2 different parts. (i.e one with message and one without)

Currently, the addTask() method solely creates and adds the task object to the list, while the addTaskWithMessage() method prints out the "task added" message, on top of adding the task using the addTask() method.