Move code to load and save task data to Storage class
The methods for loading task data from and saving to a data file are
located in the Nell class. This makes the code in Nell difficult to
read.

Let's create a new Storage class to handle these tasks. This makes the
code in Nell more readable and easier to debug.