Split parseCommand method into multiple methods
The parsing of commands in Parser is done in a single, long
parseCommand method. This makes for a long method that is difficult to
read.

Let's split the parseCommand method into two methods, one for
single-word commands and one for multi-word commands. This makes the
code in the resulting methods and Parser more readable.