Refactor command execution to use methods in Command child classes
The code for command execution comprises multiple methods in Parser.
This makes the code in Parser difficult to read, and it does not make
sense for command execution code to be within a Parser class.

Let's move the command execution code to separate Command classes, one
for each command. This makes the code more organised and easier to
read.