Modify Parser to fix bug in commands
Previously, the commands were matched with keywords sequentially.
This created an edge case when command lower on the hierarchy included
a keyword from a command higher on the hierarchy.
For instance, "find todo" would be interpreted as a todo command due
to "todo" being matched first.

To fix this, the condition for matching "COMMAND KEYWORD" is modified
to check that said keyword is at the first index of the line.