Change code in SaveFileHandler class
Change string SAVE_FILE_PATH
Removed print statements (that check if there are errors during saving/loading process) as the try-catch blocks handle it
Removed ordinal correction as that is already handled by handleInstruction in an earlier commit
loadTask now has a default switch-case, so the return null is not needed.
Add code to create new txt file and folder if not already created.