Change format of reading/writing data in Tyrone.txt
Data format of reading/writing Tyrone.txt allows for errors to arise
from the way the user writes the task descriptions (e.g. if description
contains "(by: " substring.)

Let's change the format in which data is written to and read from
Tyrone.txt to greater resemble the user input format so as to
reduce likelihood of an erronous update of TaskList from Tyrone.txt
arising from format of user-input description.