Reactor code and creating new Classes
Currently, the coding structure lacks seperate containers to group the
type of functions in the Main class. This makes the code untidy and
unorganised.

Additional classes Ui, Parser, TaskList and Storage are created.

Ui class contains of all the required print statements for the entire
program and all classes use Ui methods to print or take in input.

Parser class deals with user input and parses it into command, the
actual task, date and more. Parser mainly calls TaskList methods
accordingly after retrieving the relevant details from the input.

TaskList class stores all the Task for the session and its methods
change the current tasklist such as adding task, deleting task, marking
them and more.

Storage class is used in the initial start up of the programme as well
as before the program closes, it loads the text file information into
the program as well as save its state before the user exits.