Implement changes to ensure jar file functionality
The code does not save the tasks properly when running the jar file
generated from the previous commit. Therefore, added a new
method under the FileHandler class to ensure that the parent directory
and the text file for saving the task data exists before writing any data
to it. In the case where the file or the directory doesn't exist, they are
created accordingly.