Add find function to bot
Currently, the scope of usage remains to smaller number of task but as
we expand the usage to accomodate higher number of tasks, the need for a
find function becomes more apparent to ease the usage of finding tasks
when the list gets long.

Find function is implemented and ivoked with the command 'find' followed
by the words that the user wishes to match. The output list all matching
tasks with those keywords.