Rename command variable to input for some methods in Parser
Storing user input in variable named command could be confusing
with command in handleInput method, which refers to the first word
of the user input (list, event, etc.).

Let's rename the command variable in methods handleDeadline,
handleEvent and handleTodo to input to reduce likelihood of readers
of code getting confused.