Separate the different class concerns
Divided the funtionalities between the different classes more strictly.
Each class such as the Parser, UI, TaskHandler and Storage can now
only access the main Fenix controller class and cannot interact with
each other as freely anymore. This helps improve modularity and in
the separation of concerns as the different classes are not as tightly
coupled anymore.