Add method in "DateTimeParser" class to validate date and time entered by user
This method first checks whether the date time format in the user command (DEADLINE and EVENT) is valid and throws the corresponding exception if it's not. Next, it checks for the 5 different parameters in the date time input. They are "day", "month" and "year" for DATE and "hour" and "min" for TIME. If any of the parameters are invalid (non-integers or contain extra non-integer characters), the parse method will throw an exception and hence the input is deemed invalid.