Fix a minor bug and update methods name
Minor bug:
Initially, in the "DeleteCommand" class, after deleting the task from the list, there was an IndexOutOfBoundsException as the "showDeleteTaskMessage" method declared in the "Ui" class needed to access the index of the deleted item, to print out the message. Therefore, to resolve this issue, the message is printed first before deleting the task from the list. As such, the "tasks.deleteTask(taskId)" line is shifted down. To ensure correctness, since the message has to be printed first before the deletion of the task from the list, the size of the Arraylist is pre-decremented by 1.