Refactor "printList" method and add new method to print list of matching tasks found
The "printList" method was refactored by extracting part of it to become another method. Initially, the "printList" method contains the code for printing out messages to the terminal and the for loop that iterates through the list and prints out all the tasks in the list. After refactoring, the latter part was extracted to another method, named "printTasksInList". This not only makes the "printList" method simpler and easier to read, but also improves reusability. In fact, the newly added method, which prints out another list containing the matching tasks found based on the "find " command, uses the newly extracted method.