Fix minor bug in "Parser" class
Initially, the ExitCommand, "bye" and ListCommand, "list" both depend on the keyword (first word/string of the user command) of the user input command. As a result, even if the user enters an errorneous input command like "bye     1" or "list     hello world", the corresponding command actions will be still taken. This is not the expected behaviour of ATOM. As such, the code logic has been changed such that only the ExitCommand and the ListCommand will depend on the full user command. The other commands will still depend on the keyword of the command.